Add borders at world boundaries (#4)
Add red and white checkered bounds at top and bottom of viewport. These
bounds act as static physics objects. In order to achieve the repeating
of the red and white checkered bounds, a `Quad` was created. The quad is
wrapped with the 16px by 8px source image `border.png` which is then
repeated horizontally.