Improve handling by adding torque to turning
Update handling controls to apply torque to the vehicle, rather than
setting the angle of the vehicle directly. This resolves the issue where
turning had no effect upon the overall rotational force applied to the
vehicle. Torque is calculated assuming motion of a point particle:
`L = Iw`, where `I` is the moment of inertia and `w` is the angular
velocity. `angularVelocity` (or `w`) is calculated from `turnMultiplier`
and direction of turn, or sign of the multiplier.