Add friction between vehicle and surface
When looking for top-down physics examples, came across this example
https://love2d.org/forums/viewtopic.php?f=5&t=77596. In order to achieve
a similar effect I have repurposed most of the details of the `update`
function and performed a massive refactoring on ours.

`rotateVector` was originally `rotatePoint`, but that seemed a bit
unclear, given that this is the theorem for rotating a point
(`x`, `y`) by `angle` in two-dimensional space.

`createJoint` is used to create a new joint for use in calculating
friction on between a vehicle and a given surface.

Bump up turn multiplier from `4 -> 8`.

Remove commented code.

Draw bounding box for to aid with debugging collision.