Improve Handling (#5)
* Add speedometer

* Improve handling by adding torque to turning

Update handling controls to apply torque to the vehicle, rather than
setting the angle of the vehicle directly. This resolves the issue where
turning had no effect upon the overall rotational force applied to the
vehicle. Torque is calculated assuming motion of a point particle:
`L = Iw`, where `I` is the moment of inertia and `w` is the angular
velocity. `angularVelocity` (or `w`) is calculated from `turnMultiplier`
and direction of turn, or sign of the multiplier.

* Add friction between vehicle and surface

When looking for top-down physics examples, came across this example
https://love2d.org/forums/viewtopic.php?f=5&t=77596. In order to achieve
a similar effect I have repurposed most of the details of the `update`
function and performed a massive refactoring on ours.

`rotateVector` was originally `rotatePoint`, but that seemed a bit
unclear, given that this is the theorem for rotating a point
(`x`, `y`) by `angle` in two-dimensional space.

`createJoint` is used to create a new joint for use in calculating
friction on between a vehicle and a given surface.

Bump up turn multiplier from `4 -> 8`.

Remove commented code.

Draw bounding box for to aid with debugging collision.

* WRAP DAT WORLD 😹

* Change vehicle sprite when image dropped on game

Update player vehicle image when user drags and drops and image on the
game. No physics recalculations are done, thus the vehicle functions the
same, including the bounding box.

* Correct love compatibility version