Begin decoupling player from road
Remove `trajectory` calculation from `player:update` due to the fact
that it is only used when placing the `frontier` and road `segment`
entities. Additionally remove `lastTrajectory` and `deltaTrajectory` as
they appear to be unused.

Decouple `load` function due to the realization that it really only
needs an origin and segment dimensions. The origin `x` and `y`
coordinates are still based upon the position of the `player, but that
is now exposed in `main`. And by repurposing `width` and `height` to
correlate directly to the width and height of the road segments, we no
longer need to create a segment and measure it before we can create the
`frontier.shape`.

Since we are longer calculating `trajectory` in `player:update` this now
happens in `road:update`, where it is used to set the angle of the
`frontier`. And due to the direct relationship between the frontier and
the `segment` entities created, `addsegment` is updated to take no
incoming parameters and reference `self.frontier` instead of the
`leader`.