Move frontier collision out of road:update
Modify `road:update` to no longer detect if `player` is colliding with
the `frontier` entity. This is now in `main` along with the calculation
of `angle` (formerly `trajectory`). Update parameters of `road:update`
to only require the necessities for placing the frontier: `x`, `y`, and
`angle`. Additionally remove the `dt` parameter as it is unused.