Restructure Into Modules (#15)
* Create assets module

Clean up source directory by moving all images into `src/assets`. Create
`assets` module which loads all images used in game at once, rather than
when each entity is created.

* Move player into separate module

Create module which abstracts `draw`, `load`, and `update` callbacks of
`player`.

* Move world into local and only update when non-nil

* Remove width, and height from globals

Update `player:load` to take `world`, `width`, and `height` instead of
relying on globals to be set.

* Add entity module for ground

Move ground logic into separate module to simplify main.

* Remove references to globals from player:update

* Move road into separate module

Further simplify main by removing all road paving logic into separate
module. Rename `makeRoad -> makesegment`.

* Make world a global

Update `world` to be globally exposed to avoid passing it to each
individual entity and removing the dependency when calling `update` on
`road` entity.

* Remove wrapping when crossing window boundary

Allow player to cross window boundary, rather than "teleporting" them to
the opposite side of the screen. This removes the dependency on window
`width` and `height` when calling `player:update`.

* Scaffold out systems

* Add tiny-ecs

* Remove debug text from player and road drawing

* Remove tiny-ecs

Remove tiny-ecs library until we actually have a direct need for it.

* Begin decoupling player from road

Remove `trajectory` calculation from `player:update` due to the fact
that it is only used when placing the `frontier` and road `segment`
entities. Additionally remove `lastTrajectory` and `deltaTrajectory` as
they appear to be unused.

Decouple `load` function due to the realization that it really only
needs an origin and segment dimensions. The origin `x` and `y`
coordinates are still based upon the position of the `player, but that
is now exposed in `main`. And by repurposing `width` and `height` to
correlate directly to the width and height of the road segments, we no
longer need to create a segment and measure it before we can create the
`frontier.shape`.

Since we are longer calculating `trajectory` in `player:update` this now
happens in `road:update`, where it is used to set the angle of the
`frontier`. And due to the direct relationship between the frontier and
the `segment` entities created, `addsegment` is updated to take no
incoming parameters and reference `self.frontier` instead of the
`leader`.

* Move frontier collision out of road:update

Modify `road:update` to no longer detect if `player` is colliding with
the `frontier` entity. This is now in `main` along with the calculation
of `angle` (formerly `trajectory`). Update parameters of `road:update`
to only require the necessities for placing the frontier: `x`, `y`, and
`angle`. Additionally remove the `dt` parameter as it is unused.

* Add getTrajectory to player

* Correct reverse turning