implement global gamectx module
Also known as ctx, the global gamectx module provides a mechanism for
accessing game state in a way that can be safely shared across all
threads in the game.

The global gamectx starts a thread on load().  This thread is
responsible for storing and serializing all access to the global
gamectx's state.  Access to the state is granted via the request
channel.

To request state, a query and response channel are passed through the
request channel.  Once the management thread has processed the request,
it will send the result through the provided response channel, as if it
were a callback.  Updating state is achieved in much the same way, but a
response channel is not currently used.

The global gamectx interface handles all the channel coordination.