🔧 Use more realistic values for vehicle physics
Set `acceleration` to that of a sports car, 6.5 m/s^2. Reduce
`turnMultiplier` to prevent turning in place. Set coefficients of
friction for linear and torsional force to that of an ordinary car tire
on road. Update mass to 1350 kilograms, the average weight of a compact
car.

This does not really work right yet...