implement netman module
This is the start of some basic client/server multiplayer networking.

What works so far:

   - host a server by default
   - connect to a server with the --connect command-line argument
   - all networking is done on a dedicated thread
   - buffered sending/receiving of coordinate data over the network
   - a proto module to handle message serialization
   - an independent and configurable network tick rate
   - some basic commands for connection management

The netman module will generate regular love events for new data from
the network.  These are used to receive player data or other information
about the state of the network.  Sending data is done by calling methods
on the netman module iself, such as netman:sendCoord().

The default NET_TICK_S is only 1 (1 tick per second) so that debugging
is easier for now.  This can be set to a more realistic value when we
actually want to see stuff updated in a more real-time fashion.