implement multiplayer
This patch separates player entities from the player module.  Now, we
can call player:create() to create and return a new player entity.  The
game keeps track of a table of playersById to draw a local copy of
remote players.

The netman handlers now update the local copy of player coordinates with
fresh data from the remote side.  We also bump the proto version to add
some new commands to announce player joins/departures.

I bumped the tickrate to 30 and removed the debug messages from
netmanRecvCoord(), since we actually see other cars on the screen now.