send road data over network
This patch makes a fundamental change to the way clients interact with
the frontier.  Now, only the server can decide what influences where
the frontier goes and which cars get to push it.  The server bundles up
the road data and broadcasts it to the clients whenever it changes.
Also, we now send the entire road history to new clients in the welcome
message so that they may draw a all the road they missed out on before
connecting.

I did a little more refactoring in the road module so that we can easily
apply the received road data using a list of segment coordinates.  Also,
the main update loop is now smaller.

To support sending road data, netman now has a SEND_ROAD send type.  The
server can send road data with netman:sendRoadData().  The proto will
return freshly received road data through a new channel that netman
manages.  Clients can get the road updates via netman:recvRoadData().